+++ packages/plugin-divider/web/src/toolbar/DividerSizeButton.tsx
@@ -1,0 +1,39 @@
+import React, { useContext } from 'react';
+import type { FC, ComponentType } from 'react';
+import { RicosContext } from 'ricos-context';
+import type { IToolbarItem } from 'ricos-types';
+import { DropdownButton, ListItemSelect } from 'wix-rich-content-toolbars-ui';
+import { dividerSizeData } from './dividerButtonsData';
+
+type Props = {
+  toolbarItem: IToolbarItem;
+};
+
+export const DividerSizeButton: FC<Props> = ({ toolbarItem }) => {
+  const { t } = useContext(RicosContext) || {};
+
+  const nodeSize = toolbarItem?.attributes.nodeSize;
+  const selectedSize = (nodeSize as string) || 'CONTENT';
+  const SelectedSizeIcon = dividerSizeData.find(({ commandKey }) => commandKey === selectedSize)
+    ?.icon as ComponentType;
+
+  return (
+    <DropdownButton
+      dataHook={'baseToolbarButton_type'}
+      id={'divider_size'}
+      options={dividerSizeData.map(({ dataHook, icon: Icon, text, commandKey, tooltip }) => (
+        <ListItemSelect
+          key={commandKey}
+          dataHook={dataHook}
+          prefix={<Icon />}
+          title={t(text)}
+          selected={commandKey === selectedSize}
+          tooltip={t(tooltip)}
+          onClick={() => toolbarItem.commands?.click({ size: commandKey })}
+        />
+      ))}
+      Icon={SelectedSizeIcon}
+      tooltip={t('ButtonModal_Size_Section')}
+    />
+  );
+};